Resume index.html
I made my resume into a site. Everything looks good but for some
reason, when you view in a browser, the <div> tag stops working near
the bottom. When you get to the “Heavenly” resort tab, you’ll notice
the dates of employment are pulled in a bit. Then from there down, all
the right side elements are pulled in and when I tried to center my
footer, it gets centered from the text above it.

I would copy and paste the tags that worked above for the previous
<div>’s but for some reason it’s pulled in. Not sure why. Let me know
if you notice anything obvious…