Resume
I was able to use the CSS we learned and did some exploring of my own
on both Wades site ( www.wadehammes.com) and a talent of our’s
(www.georgemokhiber.com) and take some things to play with. Please let
me know your thoughts and things that I could cut down on to improve my
CSS and organizing.

Cheers!